% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{calc_case_control_sample_size}
\alias{calc_case_control_sample_size}
\title{计算病例对照研究的样本量并可视化}
\usage{
calc_case_control_sample_size(p1, p2, alpha = 0.05, power = 0.8)
}
\arguments{
\item{p1}{病例组成功率}

\item{p2}{对照组成功率}

\item{alpha}{显著性水平}

\item{power}{检验功效}
}
\value{
所需样本量
}
\description{
计算病例对照研究的样本量并可视化
}
