% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{calc_cross_sectional_proportion_sample_size}
\alias{calc_cross_sectional_proportion_sample_size}
\title{计算横断面调查（定性资料）的样本量并可视化}
\usage{
calc_cross_sectional_proportion_sample_size(p, alpha = 0.05)
}
\arguments{
\item{p}{预期成功率}

\item{alpha}{显著性水平}
}
\value{
所需样本量
}
\description{
计算横断面调查（定性资料）的样本量并可视化
}
