% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{calc_diagnostic_test_sample_size}
\alias{calc_diagnostic_test_sample_size}
\title{计算诊断试验的样本量并可视化}
\usage{
calc_diagnostic_test_sample_size(sensitivity, specificity, alpha = 0.05)
}
\arguments{
\item{sensitivity}{灵敏度}

\item{specificity}{特异度}

\item{alpha}{显著性水平}
}
\value{
所需样本量
}
\description{
计算诊断试验的样本量并可视化
}
