% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{calc_rct_mean_sample_size}
\alias{calc_rct_mean_sample_size}
\title{计算随机对照试验的均数比较样本量并可视化}
\usage{
calc_rct_mean_sample_size(mu1, mu2, sd, alpha = 0.05, power = 0.8)
}
\arguments{
\item{mu1}{第一组均值}

\item{mu2}{第二组均值}

\item{sd}{标准差}

\item{alpha}{显著性水平}

\item{power}{检验功效}
}
\value{
所需样本量
}
\description{
计算随机对照试验的均数比较样本量并可视化
}
